backend/requirements.txt
Django>=3.1.1<3.3
djangorestframework>=3.11.1<3.12
django-rest-auth>=0.9.5<0.10
# django-rest-registration>=0.6.0<0.7
django-cors-headers>=3.5.0<3.6
psycopg2-binary>=2.8.5<2.9
pandas>=1.1.1<1.2
gunicorn>=20.0.4<20.1
numpy>=1.19.4<1.20
Unipath>=0.2.1<0.3
django-filter>=2.3.0<2.4